Move loadSchemas from Eagle to Utils, move genPGT from Eagle to Translator.
